adjustments

+++ src/routes/index.ts
@@ -1,29 +1,29 @@
 import type { Request, Response, MethodsTypes } from 'http';
+import { compare, hash, genSalt } from 'bcrypt';
+import Prisma from '../database/prisma';
 
 type Routes = Record<string, (req: Request, res: Response) => void>;
 
-type User = {
-  name: string;
+interface User {
   email: string;
   password: string;
+}
+
+interface RegisterUser extends User {
+  name: string;
   confirmPassword: string;
-};
+}
+
+type AuthUser = User;
 
 export default (async () => {
-  const { default: Prisma } = await import('../database/prisma');
-  const { prisma } = await Prisma;
+  const {
+    default: { Headers }
+  } = await import('../utils/response');
 
-  const Headers = async (header?: Record<string, string>) => {
-    const defautHeader = { 'Content-Type': 'application/json' };
+  const { prisma } = Prisma;
 
-    if (header) {
-      return { ...defautHeader, ...header };
-    }
-
-    return defautHeader;
-  };
-
-  const verifyMethod = async (
+  const verifyMethod = (
     method: MethodsTypes,
     ctx: { req: Request; res: Response },
     callback: () => void
@@ -33,7 +33,7 @@ export default (async () => {
     if (req.method === method) {
       callback();
     } else {
-      res.writeHead(405, await Headers());
+      res.writeHead(405, Headers());
       res.json({ error: 'Method Not Allowed' });
     }
   };
@@ -41,20 +41,16 @@ export default (async () => {
   const routes: Routes = {
     '/api/users': (req: Request, res: Response) => {
       verifyMethod('GET', { req, res }, async () => {
-        const { default: cookies } = await import('../utils/cookies');
-        const { parseCookies } = await cookies;
+        console.log(req.cookies);
 
-        console.log(parseCookies(req.headers.cookies as string));
-
-        res.writeHead(200, await Headers());
+        res.writeHead(200, Headers());
         res.json(await prisma.user.findMany());
       });
     },
     '/api/register/user': (req: Request, res: Response) => {
       verifyMethod('POST', { req, res }, async () => {
-        const { hash, genSalt } = await import('bcrypt');
-
-        const { email, name, password, confirmPassword }: User = await req.body;
+        const { email, name, password, confirmPassword }: RegisterUser =
+          await req.body;
 
         const user = await prisma.user.findUnique({
           where: {
@@ -63,25 +59,31 @@ export default (async () => {
         });
 
         if (!email) {
-          res.writeHead(406, await Headers());
+          res.writeHead(406, Headers());
           res.json({ error: 'You must provide a email' });
           return;
         }
 
+        if (!email.match(/[a-z0-9._%+-]*@[a-z]*.com/i)) {
+          res.writeHead(406, Headers());
+          res.json({ msg: 'Invalid email format' });
+          return;
+        }
+
         if (user?.email) {
-          res.writeHead(406, await Headers());
+          res.writeHead(406, Headers());
           res.json({ error: 'Email already exists, try other' });
           return;
         }
 
         if (!name) {
-          res.writeHead(406, await Headers());
+          res.writeHead(406, Headers());
           res.json({ error: 'You must provide a name' });
           return;
         }
 
         if (password !== confirmPassword) {
-          res.writeHead(406, await Headers());
+          res.writeHead(406, Headers());
           res.json({ error: 'Password is not equal' });
           return;
         }
@@ -99,7 +101,7 @@ export default (async () => {
           data
         });
 
-        res.writeHead(200, await Headers());
+        res.writeHead(200, Headers());
         res.json(createdData);
       });
     },
@@ -107,9 +109,8 @@ export default (async () => {
       verifyMethod('GET', { req, res }, async () => {
         const { randomBytes } = await import('crypto');
         const { sign } = await import('jsonwebtoken');
-        const { compare } = await import('bcrypt');
 
-        const { email, password } = await req.body;
+        const { email, password }: AuthUser = await req.body;
 
         const user = await prisma.user.findUnique({
           where: {
@@ -118,13 +119,13 @@ export default (async () => {
         });
 
         if (!user?.email) {
-          res.writeHead(404, await Headers());
+          res.writeHead(404, Headers());
           res.json({ msg: `User with email ${email} not found` });
           return;
         }
 
         if (!(await compare(password, user?.password))) {
-          res.writeHead(406, await Headers());
+          res.writeHead(406, Headers());
           res.json({ msg: 'Invalid password' });
           return;
         }
@@ -143,7 +144,7 @@ export default (async () => {
           token
         };
 
-        res.writeHead(200, await Headers());
+        res.writeHead(200, Headers());
         res.json(session);
       });
     },
@@ -151,7 +152,19 @@ export default (async () => {
       if (req.method === 'PUT') {
         const { id } = req.query;
 
-        const { name }: { name: User['name'] } = await req.body;
+        const user = await prisma.user.findUnique({
+          where: {
+            id: id as string
+          }
+        });
+
+        if (!user?.id) {
+          res.writeHead(404, Headers());
+          res.json({ error: 'This user does not exist' });
+          return;
+        }
+
+        const { name }: { name: string } = await req.body;
 
         const updatedUserData = await prisma.user.update({
           select: {
@@ -165,10 +178,22 @@ export default (async () => {
           }
         });
 
-        res.writeHead(200, await Headers());
+        res.writeHead(200, Headers());
         res.json(updatedUserData);
       } else if (req.method === 'DELETE') {
         const { id } = req.query;
+
+        const user = await prisma.user.findUnique({
+          where: {
+            id: id as string
+          }
+        });
+
+        if (!user?.id) {
+          res.writeHead(404, Headers());
+          res.json({ error: 'This user does not exist' });
+          return;
+        }
 
         const deletedUser = await prisma.user.delete({
           where: {
@@ -178,13 +203,13 @@ export default (async () => {
 
         res.json(deletedUser);
       } else {
-        res.writeHead(405, await Headers());
+        res.writeHead(405, Headers());
         res.json({ error: 'Method Not Allowed' });
       }
     },
     notFound: (req: Request, res: Response) => {
-      verifyMethod('GET', { req, res }, async () => {
-        res.writeHead(404, await Headers());
+      verifyMethod('GET', { req, res }, () => {
+        res.writeHead(404, Headers());
         res.json({ error: 'Not Found' });
       });
     }
